fix: `resolve_program` function
The main problem was that call of `resolve_program` function in `Actions`
struct at `src/main.rs` was always returning shell program name regardless
of whether `--shell` argument was passed or not. That happened because in
`src/main.rs` this function was always accepting shell program name from
`config` struct.

The solution for this problem was to make `resolve_program` take editor
program name to as an argument and rewrite the logic of this function.